若干のリファクタリング
